[DSL] Refactor parsers into a subfolder
