+++ src/main/scala/Main.scala
@@ -1,4 +1,5 @@
+import temple.DSL.parser.DSLParser
+
 import scala.io.Source
-import temple.DSL.DSLParser
 
 object Main extends App {}
