+++ src/main/scala/temple/DSL/package.scala
@@ -73,5 +73,4 @@ package object DSL {
       s"$key: $tag {\n$contents\n}"
     }
   }
-
 }
